epitech_intranet/requirements.txt
Django==5.3.1
djangorestframework==3.15.2
django-cors-headers==4.3.1
Pillow==11.1.0
sqlparse==0.5.3
django-redis==5.4.0
django-otp==2.0.1
qrcode[pil]==7.4.2
django-ratelimit==4.1.0
django-haystack==3.3.0
whoosh==2.7.4
channels==4.1.0
